+++ apps/pioneer/src/app/context/index.tsx
@@ -488,8 +488,6 @@ export const PioneerProvider = ({ children }: { children: React.ReactNode }): JS
       //   'http://127.0.0.1:9001/spec/swagger.json';
       const spec =
         localStorage.getItem('pioneerUrl') ||
-        // @ts-ignore
-        import.meta.env.VITE_PIONEER_URL_SPEC ||
         'https://pioneers.dev/spec/swagger.json';
       // @ts-ignore
       console.log('spec: ', spec);