+++ packages/swapkit/tokens/src/tokenLists/Maya.ts
@@ -7,7 +7,7 @@ export const list = {
     { chain: 'BTC', identifier: 'BTC.BTC', decimals: 8 },
     { chain: 'DASH', identifier: 'DASH.DASH', decimals: 8 },
     { chain: 'MAYA', identifier: 'MAYA.CACAO', decimals: 10 },
-    { chain: 'MAYA', identifier: 'MAYA.MAYA', decimals: 6 },
+    { chain: 'MAYA', identifier: 'MAYA.MAYA', decimals: 4 },
     { chain: 'ETH', identifier: 'ETH.ETH', decimals: 18 },
     { chain: 'KUJI', identifier: 'KUJI.KUJI', decimals: 6 },
     { chain: 'KUJI', identifier: 'KUJI.USK', decimals: 6 },