+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @swapkit/toolbox-utxo
 
+## 9.0.8
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.8
+
+## 9.0.7
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.7
+
+## 9.0.6
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.6
+
 ## 9.0.5
 
 ### Patch Changes