+++ packages/wallets/metamask/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @coinmasters/wallet-metamask
 
+## 8.0.8
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@9.0.8
+  - @coinmasters/toolbox-utxo@9.0.8
+  - @coinmasters/toolbox-evm@9.0.8
+
+## 8.0.7
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@9.0.7
+  - @coinmasters/toolbox-utxo@9.0.7
+  - @coinmasters/toolbox-evm@9.0.7
+
+## 8.0.6
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@9.0.6
+  - @coinmasters/toolbox-utxo@9.0.6
+  - @coinmasters/toolbox-evm@9.0.6
+
 ## 8.0.5
 
 ### Patch Changes