+++ packages/wallets/trezor/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @coinmasters/wallet-trezor
 
+## 9.0.8
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.8
+  - @coinmasters/toolbox-utxo@9.0.8
+  - @coinmasters/toolbox-evm@9.0.8
+
+## 9.0.7
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.7
+  - @coinmasters/toolbox-utxo@9.0.7
+  - @coinmasters/toolbox-evm@9.0.7
+
+## 9.0.6
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/helpers@9.0.6
+  - @coinmasters/toolbox-utxo@9.0.6
+  - @coinmasters/toolbox-evm@9.0.6
+
 ## 9.0.5
 
 ### Patch Changes